requirements.txt
dj-database-url==0.4.1
Django==1.9
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
Pillow
django-braces
django-embed-video==1.0.0
djangorestframework==3.2.3
